refactor dao api
