+++ api_core/core/dao/__init__.py
@@ -1,1 +1,1 @@
-from .csv_load import PibDao
+from .pib_dao import PibDao
