--- api_core/core/dao/csv_load/__init__.py
@@ -1,1 +1,0 @@
-from .load_pib import PibCSVDAO as PibDao
